[sil-serialize] Improve serialization of function declarations and function definitions with external linkage
It turned out that a significant part of a swiftmodule’s size is contributed by function declarations and types used by them. Therefore it is important to avoid emitting any functions or function declarations that are not needed by the module.

This change introduces the following changes:
 - Function declarations should be serialized only if they are referenced by something that is serialized.
 - Function bodies of functions with external linkage should never be serialized.
   The only exception are shared_external functions. THis is a workaround, which will be removed in the future (see explanations in SILSerializer::writeSILFunction)

The result of the later change is that functions with external linkage are not put on the SIL serialization worklist and their bodies are not scanned, thus we do not detect references to the functions which are only referenced by functions with external linkage. In particular, the bodies of shared_external functions won’t be serialized at all if they are referenced only from functions with external linkage. And declarations of external functions referenced only from functions with external linkage won’t be serialized too.

With this changes SILSerializer emits significantly fewer function declarations.

Fixes rdar://30081040