[sil-opt] Add a new option -disable-sil-linking to dump swiftmodules "as is"
Usually, if you use `sil-opt my.swiftmodule` it would dump a lot of SIL, because it would try to link external SIL functions linked by the module (e.g. functions from the standard library) and dump them as well.

The new option allows for dumping only the SIL contained in the swiftmodule. This is very useful if you want to inspect the content of a swiftmodule.