SILGen: Add vtable output to vtable_thunks_reabstraction test
The output is rather verbose and repetitive but we want to
ensure it doesn't change unexpectedly as part of ABI stability,
and it provides a good test case for the demangler as well.

When this test was first added, concrete subclasses of generic
subclasses were not supported, but now they are, so remove some
useless generic parameters that add no value.