Sema: Narrow fix for a circular protocol conformance checking case
We go to look at the conforming protocols of an associated type
but we haven't built the generic environment for the associated
type's protocol yet.

In the test case given, we find the conformance later via a
different path, so everything continues to work.

SubstitutionMap::lookupConformance() is hopefully getting a
makeover soon, and this hack will go away.

Fixes <rdar://problem/31302713>.