[capture-promotion] Update capture-promotion for basic semantic-sil usage.
In this commit I update capture-promotion for very basic usage and verify this
by adding a version of one of the tests with ownership (and thus the verifier)
enabled.

I am going to move over the rest of the tests in subsequent commits with changes
necessary for capture promotion to work with them.

rdar://29870610