Enable The Optional Sanitizer (by default)
The Optional Sanitizer is a revolutionary new tool that allows
developers to gain enhanced insight into the inner workings of their
programs.  Traditional debugging techniques rely on a boring terminal
with text scrolling by, and a thousand commands that you need to
remember.  Instead, we have simplified the process of debugging down
to, what we see as, the most unused sense when programming: The ears.

That’s right, now you can hear your program as it executes! No longer
will you have to struggle with the debugger, now every optional
force-unwrap can grate on your senses with the ease and simplicity that
only Apple can provide.

Tooling is also highly accessible.  To those that wish to get all the
benefits of OSan, but may not be able to the aural sensation to its
fullest extent, we provide a visual OSan that is enabled by passing
-Xfrontend -visual-osan-unwrap-instrumentation