AST: Stricter contract for Type::subst()
Fix an odd corner case when UseErrorTypes was off; we would
return the empty type if dependent member type substitution
failed, but otherwise return the original type if it was a
generic type parameter or an archetype.

Now, if UseErrorTypes is off, return the empty type in both
cases, even if the original type is 'primary'.