[ast-verifier] Verify that inout_to_pointer and array_to_pointer are (almost) immediate children of an ApplyExpr argument.
Specifically we only allow it as an argument to a TupleExpr and a ParenExpr and
can have the following forms:

  (inject_into_optional (*_to_pointer ...))

or

  (*_to_pointer ...)

This will allow for inout_to_pointer and array_to_pointer to be pattern matched
robustly and then special cased in SILGenApply.

The reason why these expressions need to be special cased is that they can
perform inout_expr accesses while going through code paths in SILGenApply where
it is expected that inout accesses do not occur.

rdar://29870610