[Constraint system] More or less rewrite the constraint propagation pass.
This is still disabled by default, and as with the previous
implementation does not pass all tests if enabled.

Rather than looking for applicable function constraints and then looking
at the bind overload disjunctions associated with those, instead
directly collect all bind overload disjunctions and revise the enabled
elements of those disjunctions by attempting to bind them simultaneously
with the elements of other disjunctions that are linked to this one
through applicable function constraints.