[Serialization] Look through unexpected typealiases.
When an Objective-C type is renamed, the importer leaves behind a
typealias marked "unavailable". Look through that when we're expecting
a nominal type or bound generic type.

There's still more work to do here: in the case of a typealias /not/
left behind by the importer, there's a chance the generic arguments
of the typealias aren't just directly forwarded to the underlying
nominal. The right thing to do here is to build a substitution map
and call subst() on the underlying type.